Cambiado tipo de relacion ya que podria presentar errores, Modificacion en el formulario para cambiar contraseña, plantilla y Mail para cambiar la contraseña, tabla para los resets custom creada, agregada relacion de miembro a usuario

+++ database/migrations/2024_10_15_134415_create_envelopes_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('envelopes', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedBigInteger('id_member');
+            $table->unsignedInteger('id_member');
             $table->string('envelope_number');
             $table->date('date');
             $table->string('description')->nullable();
