+++ database/migrations/2024_10_15_134444_create_offerings_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('offerings', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedBigInteger('id_envelope');
-            $table->unsignedBigInteger('id_offering_type')->nullable();
+            $table->unsignedInteger('id_envelope');
+            $table->unsignedInteger('id_offering_type')->nullable();
             $table->decimal('amount', 10,2);
             $table->timestamps();
